+++ src/scorer/feature/author_followers_scorer.ts
@@ -1,12 +1,12 @@
 /*
  * Score a toot based on how many followers the author has.
  */
-import FeatureScorer from "../feature_scorer";
 import Toot from "../../api/objects/toot";
+import TootScorer from "../feature_scorer";
 import { ScoreName } from '../../enums';
 
 
-export default class AuthorFollowersScorer extends FeatureScorer {
+export default class AuthorFollowersScorer extends TootScorer {
     description = "Favour accounts with a lot of followers";
 
     constructor() {