+++ src/scorer/feature/favourited_tags_scorer.ts
@@ -1,15 +1,15 @@
 /*
  * Score how many times the current user has favourited the toot's hashtags in the past.
  */
-import FeatureScorer from '../feature_scorer';
 import TagList from '../../api/tag_list';
 import Toot from '../../api/objects/toot';
+import TootScorer from '../feature_scorer';
 import { ScoreName } from '../../enums';
 import { sumArray } from '../../helpers/collection_helpers';
 import { type StringNumberDict } from '../../types';
 
 
-export default class FavouritedTagsScorer extends FeatureScorer {
+export default class FavouritedTagsScorer extends TootScorer {
     description = "Favour toots containing hashtags you favourite";
 
     constructor() {