+++ src/scorer/feature/followed_accounts_scorer.ts
@@ -3,14 +3,14 @@
  * RetootsInFeedScorer).
  */
 import Account from '../../api/objects/account';
-import FeatureScorer from '../feature_scorer';
 import MastoApi from '../../api/api';
 import Toot from '../../api/objects/toot';
+import TootScorer from '../feature_scorer';
 import { ScoreName } from '../../enums';
 import { type StringNumberDict } from '../../types';
 
 
-export default class FollowedAccountsScorer extends FeatureScorer {
+export default class FollowedAccountsScorer extends TootScorer {
     description = "Favour accounts you follow";
 
     constructor() {