+++ src/scorer/feature/followed_tags_scorer.ts
@@ -2,12 +2,12 @@
  * Populate the 'followedTags' property on Toot object and return the number of tags
  * on the Toot that the user is following.
  */
-import FeatureScorer from '../feature_scorer';
 import Toot from '../../api/objects/toot';
+import TootScorer from '../feature_scorer';
 import { ScoreName } from '../../enums';
 
 
-export default class FollowedTagsScorer extends FeatureScorer {
+export default class FollowedTagsScorer extends TootScorer {
     description = "Favour toots containing hashtags you follow";
 
     constructor() {