+++ src/scorer/feature/num_replies_scorer.ts
@@ -1,12 +1,12 @@
 /*
  * Score how many times the toot has been replied to by other users.
  */
-import FeatureScorer from '../feature_scorer';
+import TootScorer from '../feature_scorer';
 import Toot from '../../api/objects/toot';
 import { ScoreName } from '../../enums';
 
 
-export default class NumRepliesScorer extends FeatureScorer {
+export default class NumRepliesScorer extends TootScorer {
     description = "Favour toots with lots of replies";
 
     constructor() {