+++ src/scorer/feature/video_attachment_scorer.ts
@@ -1,12 +1,12 @@
 /*
  * Score the number of video attachments in the toot.
  */
-import FeatureScorer from '../feature_scorer';
 import Toot from '../../api/objects/toot';
+import TootScorer from '../feature_scorer';
 import { ScoreName } from '../../enums';
 
 
-export default class VideoAttachmentScorer extends FeatureScorer {
+export default class VideoAttachmentScorer extends TootScorer {
     description = "Favour video attachments";
 
     constructor() {